requirements.txt
langchain_openai 
langchain_core
python-dotenv
streamlit
langchain_community
langchain_ollama
fastapi
uvicorn
sse_starlette
bs4
pypdf
chromadb
faiss-cpu
langchain-groq